backend/baymax/requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain_pinecone
langchain_community
langchain_openai
langchain_ollama
langchain_experimental
fastapi
uvicorn[standard]
motor
-e .